+++ src/main/java/com/xdf/xd_f371/entity/Inventory.java
@@ -19,21 +19,21 @@ public class Inventory {
     @Column(name = "petro_id")
     private int petro_id;
     @Column(name = "tdk_nvdx")
-    private int tdk_nvdx=0;
+    private double tdk_nvdx=0;
     @Column(name = "tdk_sscd")
-    private int tdk_sscd=0;
+    private double tdk_sscd=0;
     @Column(name = "nhap_nvdx")
-    private int nhap_nvdx=0;
+    private double nhap_nvdx=0;
     @Column(name = "nhap_sscd")
-    private int nhap_sscd=0;
+    private double nhap_sscd=0;
     @Column(name = "xuat_nvdx")
-    private int xuat_nvdx=0;
+    private double xuat_nvdx=0;
     @Column(name = "xuat_sscd")
-    private int xuat_sscd=0;
+    private double xuat_sscd=0;
     @Column(name = "status")
     private String status;
     @Column(name = "price")
-    private int price=0;
+    private double price=0;
     @Column(name = "create_at", insertable = false,updatable = false)
     private LocalDate create_at;
     @Column(name = "sd")
@@ -47,7 +47,7 @@ public class Inventory {
     @JoinColumn(name = "petro_id",referencedColumnName = "id", insertable = false, updatable = false)
     private LoaiXangDau loaiXangDau;
 
-    public Inventory(int petro_id, int tdk_nvdx, int tdk_sscd, int nhap_nvdx, int nhap_sscd, int xuat_nvdx, int xuat_sscd, String status, int price,
+    public Inventory(int petro_id, double tdk_nvdx, double tdk_sscd, double nhap_nvdx, double nhap_sscd, double xuat_nvdx, double xuat_sscd, String status, double price,
                      LocalDate sd, LocalDate ed) {
         this.petro_id = petro_id;
         this.tdk_nvdx = tdk_nvdx;
